[crypto] Unplug ValidSignatureBytesBLST
The blst keygen is unstable on 1.15 (in ways that are hard to reproduce), leading to CI failures. https://github.com/onflow/flow-go/pull/1119 disabled the use of the blst keygen in generating valid slices of scalar bytes for cross-testing bls.go with BLST, this completes the job in de-activating the use of the keygen for creating siganture material.
We leave the auxiliary functions in place for future upgrades (the blst keygen is stable on go 1.16 and some architectures on go 1.15)